requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncio==3.4.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.5.0
google-auth==2.39.0
google-genai==1.10.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
multidict==6.4.3
pillow==11.2.1
propcache==0.3.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyNaCl==1.5.0
python-dotenv==1.1.0
python-nasa-api==0.0.2
pytz==2025.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
websockets==15.0.1
yarl==1.19.0
yt-dlp==2025.3.31
